更新StringUtils及部分使用场景

+++ abc-modules/abc-system/src/main/java/com/abc/system/service/impl/SysDictTypeServiceImpl.java
@@ -213,9 +213,9 @@ public class SysDictTypeServiceImpl implements ISysDictTypeService
     @Override
     public String checkDictTypeUnique(SysDictType dict)
     {
-        Long dictId = StringUtils.isNull(dict.getDictId()) ? -1L : dict.getDictId();
+        long dictId = StringUtils.isNull(dict.getDictId()) ? -1L : dict.getDictId();
         SysDictType dictType = dictTypeMapper.checkDictTypeUnique(dict.getDictType());
-        if (StringUtils.isNotNull(dictType) && dictType.getDictId().longValue() != dictId.longValue())
+        if (StringUtils.isNotNull(dictType) && dictType.getDictId() != dictId)
         {
             return UserConstants.NOT_UNIQUE;
         }
